instedd/cepheid-web#859: Serialize requests to FieldsService to avoid race condition
We are currently performing a second request to the fields service to get the label for the current assay.
Until now, we were assuming the callback for the second request was called after the first one, which was not necessarily true.

Quick fix: serialize the requests so that the second one is perfomed only only after the first one.

We should probably find a way to avoid this second request completely (maybe including the assay label in the report object)